Ladders & Jumping
Added ladder tiles that work

Fixed the jumping so that edges of the player can be off a ledge and they can still jump
