+++ Prototype/Assets/Prefabs/Ladders.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ead916380554cd14d919ceac47810083
+timeCreated: 1516494149
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
