+++ Prototype/Assets/Scripts/RatController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 13cc65d1ba9cb5e42beeb55e207a9ed2
-timeCreated: 1516313411
+guid: 512a719917b841f458d82d48080a08d5
+timeCreated: 1516490796
 licenseType: Free
 MonoImporter:
   externalObjects: {}
